Patterns: Use CopyPatternButton in thumbnail
This consolidates the copy code, removing the duplicate code from PatternThumbnail
