+++ public_html/wp-content/themes/pattern-directory/src/components/pattern-thumbnail/index.js
@@ -1,46 +1,16 @@
 /**
  * WordPress dependencies
  */
-import { speak } from '@wordpress/a11y';
-import { __, sprintf } from '@wordpress/i18n';
-import { Button, Disabled } from '@wordpress/components';
-import { useEffect, useState } from '@wordpress/element';
+import { Disabled } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import FavoriteButton from '../favorite-button';
+import CopyPatternButton from '../copy-pattern-button';
 import Canvas from './canvas';
-import { copyToClipboard } from '../../utils';
 
 function PatternThumbnail( { pattern } ) {
-	const [ copied, setCopied ] = useState( false );
-
-	const handleCopy = () => {
-		const result = copyToClipboard( pattern.pattern_content );
-
-		setCopied( result );
-	};
-
-	useEffect( () => {
-		if ( ! copied ) {
-			return;
-		}
-
-		speak(
-			sprintf(
-				/* translators: %s: pattern title. */
-				__( 'Copied %s pattern to clipboard.', 'wporg-patterns' ),
-				pattern.title.rendered
-			)
-		);
-
-		const timer = setTimeout( () => setCopied( false ), 20000 );
-		return () => {
-			clearTimeout( timer );
-		};
-	}, [ copied ] );
-
 	return (
 		<div className="pattern-grid__pattern">
 			<a href={ pattern.link } rel="bookmark">
@@ -52,9 +22,7 @@ function PatternThumbnail( { pattern } ) {
 			<div className="pattern-grid__actions">
 				<h2 className="pattern-grid__title">{ pattern.title.rendered }</h2>
 				<FavoriteButton showLabel={ false } patternId={ pattern.id } />
-				<Button className="pattern__copy-button is-small" isPrimary onClick={ handleCopy }>
-					{ copied ? __( 'Copied!', 'wporg-patterns' ) : __( 'Copy', 'wporg-patterns' ) }
-				</Button>
+				<CopyPatternButton isSmall={ true } content={ pattern.pattern_content } />
 			</div>
 		</div>
 	);
